правки 2.32

+++ internal/models/models.go
@@ -25,7 +25,7 @@ type Order struct {
 	ID         int        `json:"id,omitempty"`
 	Number     string     `json:"number,omitempty"`
 	Accrual    float64    `json:"accrual"`
-	Withdrawn  int        `json:"sum"`
+	Withdrawn  float64    `json:"sum"`
 	Status     string     `json:"status,omitempty"`
 	UploadedAt *time.Time `json:"uploaded_at,omitempty"`
 	UserID     int        `json:"user_id,omitempty"`
